Revert "Koren karta dodata"
This reverts commit 39871f612ba84d661b5a4898b4d6ef563c638726.

+++ Assets/Skripte/AdditionCard.cs
@@ -4,13 +4,13 @@ using UnityEngine;
 
 public class AdditionCard : AbstractCard {
 
-    public double value = 1;
+    public float value = 1;
 
     public override void Apply (GameManager manager, Player player) {
         player.Lifepoints += value;
     }
 
-    public void SetValue (double newValue) {
+    public void SetValue (int newValue) {
         value = newValue;
 
         Description = $"+{value}";
