+++ Assets/Skripte/MnozenjeCard.cs
@@ -5,14 +5,14 @@ using UnityEngine;
 public class MnozenjeCard : AbstractCard
 {
 
-    public double value = 1;
+    public float value = 1;
 
     public override void Apply(GameManager manager, Player player)
     {
         player.Lifepoints *= value;
     }
 
-    public void SetValue(double newValue)
+    public void SetValue(int newValue)
     {
         value = newValue;
